Allow action menus to specify order and add separators.
Menu-type actions were restricted to rendering in the order that actions
were registered, but this is highly annoying when pulling in actions
from multiple different places, including extensions.

This change adds the ability for a menu action to define the order of
the children, and adds a new separator implementation (defined with the
custom "--" name in the order). Any actions which are in the menu but
not explicitly specified in the child order will be added at the end of
the menu.

Testing Done:
Created a menu action and specified the child order. Saw that I was able
to reorder the menu items as expected. Verified that separators worked
and looked good.

Reviewed at https://reviews.reviewboard.org/r/13162/