Fix publishing a review request and review at the same time.
I seem to have made a mistake with the code that enabled publishing a
review request and the review at the same time. I added all the right
plumbing but apparently never actually hooked it up. This change does
that, and adds some relevant unit tests to ensure that this works as
expected.

Testing Done:
- Verified that I could publish a (non-public) review request along with
  a review at the same time.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13190/