Split out all hosting service base support into new modules.
Our hosting service support was introduced fairly early in the life of
Review Board, and the base support has outgrown the scattered files
they've lived in. With some upcoming work on the horizon to both remove
old code and to add new support for SSL certificates and eventually
deeper bug tracker integration, it's time to clean up a lot of this.

This change splits `reviewboard.hostingsvcs.service` into a series of
modules in `reviewboard.hostingsvcs.base`, as well as moving
`reviewboard.hostingsvcs.utils.paginator` and
`reviewboard.hostingsvcs.{forms,repository}` into there.

To keep this reviewable, there are no changes to these files aside from
some "Version Changed" docs in classes detailing the move, and new
module docstrings. Currently, nothing is marked as deprecated with a
warning, and all prior imports will continue to work.

Follow-up changes will rename some of the new classes to better signify
base classes and to add typing. For Review Board 6, the old locations
will not be deprecated, leaving that to Review Board 7. This is to avoid
introducing major deprecations this late in the development cycle.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13175/