Modernize reviewboard.hostingsvcs.base.http and add typing.
This updates the typing for the hosting service HTTP support, updating
some legacy documentation and bad assumptions in the code accordingly.
This includes the removal of a lot of Python 2-isms in the code and
documentation that are no longer needed.

Functionality marked as deprecated in Review Board 4 have now been given
a deprecation warning, scheduled for removal in Review Board 7.

Testing Done:
Unit tests pass.

Verified that pyright and mypy were happy with this module.

Reviewed at https://reviews.reviewboard.org/r/13177/