Fix a bug with showing content of deleted files in the diff viewer.
This fixes a bug with the show content link for deleted files in the diff
viewer. The link wouldn't show the content of the file and would instead
remove it from the diff viewer. This is fixed by properly replacing existing
file diffs in the diff viewer when the show content link is clicked.

Testing Done:
- Clicked on show content for multiple files in a diff viewer and saw that
  they were shown properly.
- Saw that the diff file index properly listed files depending on which
  files are visible on screen, both before and after clicking show content.
- Saw that diffs were properly rendered for non-deleted files.
- Ran JS unit tests.

Reviewed at https://reviews.reviewboard.org/r/13196/