Properly type SCMTool.scmtool_id as a string.
We define `SCMTool.scmtool_id` as a helper property that computes an ID
if not already set, based on an Entrypoint. This is here because we used
to use entrypoint names as the IDs, and while transitioning away from
that, we needed to maintain compatibility.

We still need this compatibility shim (whether here or in registration,
but at the moment it's here). Power Pack still fails to set an SCMTool
ID for TFS and TFS-Git, so this is still relevant.

Because we assign this to a property, type checkers think the proper
type is that of the helper. This change explicitly marks it as a `str`.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13200/