Broaden and rework the Review Board color palette.
This extends our color palette to give us a broader set of colors we can
use on the page for light mode.

When the original color palette was created, we only had the colors we
needed at the time, leaving gaps in the palette. As time went on, some
new colors were added to fill that in, but it left some of the color
gradients a bit inconsistent.

Now, each color type has been filled with a consistent gradient of
colors. This required minimal adjustment, except to a red color usage.

Some new color categories have been created: Cool Grey, Mustard
(replacing the old Brown), a new Brown, and Cyan.

Most of these are not used, but the styling update for 6.0 will be
making use of some of these new definitions.

Testing Done:
Tested the styles that used the affected constants. Verified that,
while there are color differences when you pixel-peep, they still
look the same in context and don't appear to have any visual issues.

Reviewed at https://reviews.reviewboard.org/r/13205/