Fix overzealous issue summary table reloads.
The review request page does some long-running requests to fetch new
updates to various parts, including entries and the issue summary table.
The entries all had an updated timestamp included, and on the client
side we'd check to make sure there actually was an update, but the issue
summary table did not. This change adds a new timestamp to the update
payload for this, and checks that timestamp client-side to prevent us
from doing unnecessary DOM updates.

Testing Done:
- Ran python tests.
- Ran js-tests.
- Verified that when just looking at an unchanged page, the issue
  summary table was not being updated in the DOM.
- Changed some issue state from another tab and saw the issue summary
  table reload correctly.

Reviewed at https://reviews.reviewboard.org/r/13234/