Catch Bazaar/Breezy import errors and show them.
If `bzr`/`brz` fail to start due to an import error (such as a missing
Paramiko or another dependency), the error would be masked behind a
"Repository not found" error.

We now detect this and represent the error as an `SCMError`, helping to
debug this both in production and during test runs.

Testing Done:
Used this to debug some unit test failures. Verified that the exception
was represented properly.

Reviewed at https://reviews.reviewboard.org/r/13229/