Add documentation throughout the Subversion code.
This fleshes out our codebase documentation for Subversion, documenting
all modules, functions, and classes, in an effort to clean up the code
and explain the purposes of some functions in preparation for upcoming
code cleanups needed for the SSL functionality.

Some of the docs reference types that don't currently exist, but will be
available in the next change in this series.

Testing Done:
Built the docs and checked for any new errors.

Reviewed at https://reviews.reviewboard.org/r/13224/