Use the modern spinner throughout the UI.
In Review Board 5, we introduced a new SVG-based spinner, which lacked
the wobble of the Font Awesome glyph-based spinner. All of Djblets and
parts of Review Board were updated to use it, but we still had plenty of
code using the old spinner.

This change converts the rest of the codebase to use the new one. It's
basically a drop-in replacement, though it did require some small
adjustments in some parts of the UI to better fit the size or visuals.
In particular, file attachment thumbnails now use margins to position
it, and, given the spinner size, reduce the opacity to make it less
strong.

Testing Done:
Tested each spinner in the UI, before and after the change, to make
sure that sizes were still correct.

Reviewed at https://reviews.reviewboard.org/r/13232/