Fix a regression when accepting a SSL certificate.
The new client-building code in Subversion returns a client class and
not a tuple. One of the calls to this, used when accepting a SSL
certificate, was still treating it as a tuple, causing a crash when
accepting the certificate.

This is a small fix for updating the call site. Note that this function
will be removed when the new SSL code is in.

Testing Done:
Verified that type checking was happy.

Reviewed at https://reviews.reviewboard.org/r/13243/