Mark the package README as ReStructuredText.
Historically, the README for a package would be handled as
ReStructuredText by default, and this is still documented as the case.
However, modern packaging tools seem to be marking this as plain text by
default, impacting our listings on PyPI.

This is a trivial change that marks the description as ReStructuredText.