Move the '#' character out of the link on the review request box.
While this is not completely consistent across browsers, Chrome allows
people to right click on a link and select "copy", which will copy the
text of the link. We had a user report that it was awkward using this
with the ID in the review request box because it would include a '#'
character, making it so that pasting into other systems required manual
editing.

In order to facilitate copying just the review request ID number, this
change moves the '#' out of the link element.

Testing Done:
In Chrome, right clicked on the review request ID link and copied it.
Pasted and saw that the clipboard only contained the ID number without
the hash character.

Reviewed at https://reviews.reviewboard.org/r/13269/