Introduce file attachment states.
Our file attachments can be in many different states, but we don't really have
a solid concept of these states in our codebase so far. We also don't have a
formal way of determining which state a file attachment is in.

This change introduces file attachment states and allows us to efficiently
get the state of a file attachment through its review request or draft.

This sets the foundation for an upcoming change that improves how we display
file attachments on a review request and how users can manage them.

Testing Done:
- Ran unit tests.
- Ran JS unit tests.
- Used in upcoming changes.

Reviewed at https://reviews.reviewboard.org/r/13238/